website/requirements/django-1.8-upgrade.txt
# upgrade
django-threadedcomments
django-fluent-comments==1.0.4
django-comments==0.3.3a0
django-filter==0.11.0
django-polymorphic==0.7.2
django-reversion==1.9.3
django-mptt==0.7.4
django-filer==0.9.12


django-cms==3.1.2
django-treebeard==3.0
djangocms-admin-style==0.2.8
djangocms-file==0.1
djangocms-link==1.6.2
djangocms-picture==0.1
djangocms-snippet==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.6.0
djangocms-video==0.1
cmsplugin-filer==0.10.2

django-cache-machine==0.9

django-sekizai==0.8.2

django-celery==3.1.16
celery==3.1.16
django-guardian==1.3
django-hvad==1.3.0

django-extensions==1.5.7
django-dajaxice==0.7

# uninstall
pip uninstall django-modeltranslation
pip uninstall south




# migration steps

./manage.py migrate --fake-initial
./manage.py migrate cms 0001 --fake
./manage.py migrate cms 0002 --fake

